ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pillow==11.1.0
python-multipart==0.0.20
numpy==2.2.2
tensorflow==2.18.0
